+++ app/assets/javascripts/discourse/app/components/user-field.hbs
@@ -1,0 +1,6 @@
+<this.userFieldComponent
+  @field={{@field}}
+  @value={{@value}}
+  @class={{@class}}
+  @validation={{@validation}}
+/>
