+++ app/assets/javascripts/discourse/app/components/user-fields/confirm.hbs
